# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.CreateClientRequest:
    properties:
      credit_limit:
        minimum: 0
        type: number
      email:
        type: string
      is_active:
        type: boolean
      phone:
        type: string
      user_id:
        type: integer
    required:
    - credit_limit
    - email
    - phone
    - user_id
    type: object
...
  request.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        description: Puedes usar enum si lo prefieres
        enum:
        - user
        - client
        - admin
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
  request.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  request.ResetPasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  request.UpdateClientRequest:
    properties:
      credit_limit:
        minimum: 0
        type: number
      email:
        type: string
      is_active:
        type: boolean
      phone:
        type: string
    required:
    - credit_limit
    - email
    - phone
    type: object
...
  response.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  response.ClientResponse:
    properties:
      created_at:
        type: string
      credit_limit:
        type: number
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      phone:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing finances in small businesses.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Final Assignment Finance API Rest
  version: "1.0"
paths:
...
  /api/v1/clients:
    get:
      consumes:
      - application/json
      description: Gets a list of all clients.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ClientResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all clients
      tags:
      - Clients
    post:
      consumes:
      - application/json
      description: Creates a new client with the provided data.
      parameters:
      - description: Client data
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/request.CreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ClientResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new client
      tags:
      - Clients
...
  /api/v1/clients/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a client by its ID.
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Client deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid client ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Client not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete client
      tags:
      - Clients
    get:
      consumes:
      - application/json
      description: Gets a client by its ID.
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ClientResponse'
        "400":
          description: Invalid client ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Client not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get client by ID
      tags:
      - Clients
    put:
      consumes:
      - application/json
      description: Updates a client's data.
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated client data
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/request.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Client updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid client ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Client not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update client
      tags:
      - Clients
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Logs in with user credentials
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Authentication
...
  /api/v1/refresh:
    post:
      consumes:
      - application/json
      description: Generates a new access token from a valid refresh token.
      parameters:
      - description: Bearer {refreshToken}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuthResponse'
        "401":
          description: Invalid or expired refresh token.
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - Authentication
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the provided data
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Authentication
  /api/v1/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the password for the authenticated user, requiring the current
        password
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Reset password request
        in: body
        name: reset
        required: true
        schema:
          $ref: '#/definitions/request.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - Authentication
...
swagger: "2.0"