# internal/model/dto/request/create_user_request.go
package request

type CreateUserRequest struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=6"`
	Name     string `json:"name" binding:"required"`
	Role     string `json:"role" binding:"required,oneof=user client admin"` // Puedes usar enum si lo prefieres
}